Delete old monitoring script
